invoice details update

+++ src/main/java/com/its/service/projection/InvoiceTotalAmountProjection.java
@@ -1,5 +1,7 @@
 package com.its.service.projection;
 
+import java.math.BigDecimal;
+
 public interface InvoiceTotalAmountProjection {
-    Long getTotalPurchase();
+    BigDecimal getTotalPurchase();
 }
